test: add shared @wso2/test-config Jest preset and migrate diagram packages onto it
