+++ packages/ballerina-visualizer/package.json
@@ -75,6 +75,7 @@
     "katex": "0.16.27"
   },
   "devDependencies": {
+    "@wso2/test-config": "workspace:*",
     "jest": "29.7.0",
     "ts-jest": "29.3.4",
     "@types/jest": "29.5.14",
