--- packages/ballerina-visualizer/src/test/jest.env.ts
@@ -1,1 +1,0 @@
-(globalThis as any).structuredClone = (val: any) => JSON.parse(JSON.stringify(val));
