+++ packages/bi-diagram/jest.config.js
@@ -16,46 +16,14 @@
  * under the License.
  */
 
+const base = require('@wso2/test-config/jest-preset');
+
 module.exports = {
-    preset: 'ts-jest/presets/js-with-ts',
-    testEnvironment: 'jsdom',
-    transform: {
-        '^.+\\.(js|jsx)$': 'babel-jest',
-        '^.+\\.(ts|tsx)$': 'ts-jest',
-    },
-    globals: {
-        'ts-jest': {
-            isolatedModules: true,
-            tsconfig: {
-                jsx: 'react',
-                esModuleInterop: true,
-                allowSyntheticDefaultImports: true,
-            }
-        },
-    },
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-    testMatch: ['**/?(*.)+(spec|test).ts?(x)'],
+    ...base,
+    rootDir: '.',
     moduleNameMapper: {
-        '^@vscode/codicons/dist/codicon.css$': '<rootDir>/node_modules/identity-obj-proxy',
-        '\\.(css|less|sass|scss)$': '<rootDir>/node_modules/identity-obj-proxy',
-        '\\.(svg|png|jpg|jpeg|gif)$': '<rootDir>/node_modules/identity-obj-proxy',
-        '^react$': '<rootDir>/node_modules/react/index.js',
-        '^react-dom$': '<rootDir>/node_modules/react-dom/index.js',
+        ...base.moduleNameMapper,
         '^http-proxy-agent$': '<rootDir>/src/test/mocks/proxyAgent.ts',
         '^https-proxy-agent$': '<rootDir>/src/test/mocks/proxyAgent.ts',
     },
-    setupFilesAfterEnv: [
-        '<rootDir>/src/test/jest.env.ts',
-    ],
-    setupFiles: ['<rootDir>/src/test/matchMedia.ts'],
-    transformIgnorePatterns: [
-        '<rootDir>/node_modules/(?!(@wso2)/)' // Only transform @wso2 packages
-    ],
-    collectCoverageFrom: [
-        'src/**/*.{ts,tsx}',
-        '!src/**/*.d.ts',
-        '!src/**/*.stories.{ts,tsx}',
-        '!src/test/**/*'
-    ]
 };
-
