+++ packages/bi-diagram/package.json
@@ -51,6 +51,7 @@
         "@testing-library/dom": "10.4.0",
         "@testing-library/jest-dom": "6.6.3",
         "@testing-library/react": "16.3.0",
+        "@wso2/test-config": "workspace:*",
         "@types/dagre": "0.7.52",
         "@types/jest": "29.5.14",
         "@types/lodash": "4.17.16",
