--- packages/sequence-diagram/src/test/jest.env.ts
@@ -1,79 +1,0 @@
-/**
- * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import '@testing-library/jest-dom';
-
-(globalThis as any).structuredClone = (val: any) => {
-    return JSON.parse(JSON.stringify(val))
-}
-
-(globalThis as any).setImmediate = (globalThis as any).setImmediate || ((fn: any, ...args: any) => globalThis.setTimeout(fn, 0, ...args)) as any;
-
-class MockResizeObserver {
-  observe = jest.fn();
-  unobserve = jest.fn();
-  disconnect = jest.fn();
-}
-
-(globalThis as any).ResizeObserver = MockResizeObserver;
-
-jest.mock('resize-observer-polyfill', () => {
-  return {
-    __esModule: true,
-    default: MockResizeObserver,
-  };
-});
-
-// Mock getBoundingClientRect for canvas elements
-Element.prototype.getBoundingClientRect = jest.fn(() => ({
-  width: 800,
-  height: 600,
-  top: 0,
-  left: 0,
-  bottom: 600,
-  right: 800,
-  x: 0,
-  y: 0,
-  toJSON: jest.fn()
-}));
-
-// Mock canvas context
-HTMLCanvasElement.prototype.getContext = jest.fn(() => ({
-  fillRect: jest.fn(),
-  clearRect: jest.fn(),
-  getImageData: jest.fn(),
-  putImageData: jest.fn(),
-  createImageData: jest.fn(),
-  setTransform: jest.fn(),
-  drawImage: jest.fn(),
-  save: jest.fn(),
-  restore: jest.fn(),
-  beginPath: jest.fn(),
-  moveTo: jest.fn(),
-  lineTo: jest.fn(),
-  closePath: jest.fn(),
-  stroke: jest.fn(),
-  fill: jest.fn(),
-  scale: jest.fn(),
-  rotate: jest.fn(),
-  translate: jest.fn(),
-  measureText: jest.fn(() => ({ width: 10 })),
-  fillStyle: '',
-  strokeStyle: '',
-  globalAlpha: 1,
-})) as any;
