+++ packages/test-config/matchMedia.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,6 +15,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
+// Runs before the test framework is installed (setupFiles).
+// jsdom does not implement matchMedia or scrollIntoView.
 
 Object.defineProperty(window, 'matchMedia', {
     writable: true,
