+++ packages/type-diagram/package.json
@@ -50,6 +50,7 @@
 		"@testing-library/dom": "10.4.0",
 		"@testing-library/jest-dom": "6.6.3",
 		"@testing-library/react": "16.3.0",
+		"@wso2/test-config": "workspace:*",
 		"@types/jest": "29.5.14",
 		"@types/react": "18.2.0",
 		"@types/react-dom": "18.2.0",
